requirements.txt
requests
python-dotenv
elevenlabs